30-04

+++ Game/CreatureCommand.cs
@@ -8,11 +8,11 @@ namespace Game
 {
     class CreatureCommand
     {
-        public double DeltaX;
-        public double DeltaY;
+        public int DeltaX;
+        public int DeltaY;
         public ICreature TransformTo;
         
-        public static CreatureCommand operator * (CreatureCommand command, double a)
+        public static CreatureCommand operator * (CreatureCommand command, int a)
         {
             return new CreatureCommand { DeltaX = command.DeltaX * a, DeltaY = command.DeltaY * a, TransformTo = command.TransformTo };
         }
